Improve the "What's new" file with tables
Display some of the replacements to do in tables so it reads nicely in
plain file and it's better displayed on the site.

Some of them are still under code blocks because we lose the readability
of tables if we put them into it since the lines are very long.