Add a note ➡ config:convert
I think we need to draw more attention to this feature - if you've written your config file, the only place where config:convert is mentioned is in Building a project.  I read through this document and as I was working with an existing database,  read through 'the easy way'

I suggest that  as this is such an important aspect of the configuration, that this section be moved to it's own file, so that it can be referenced from the build docs here, and also from the configuration section - which  is also missing this information.